requirements.txt
# requirements.txt

# PostgreSQL adapter for Python Neon
psycopg2-binary

# Load environment variables from .env files
python-dotenv

# For Jupyter Notebooks (if needed)
ipykernel

# ODBC adapter for Python MS SQL Server
pyodbc